# docs-site/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SpeakerWeave API",
  "version": "1.0.0",
  "description": "Organization-scoped REST API for conference program data. Authenticate every request with an API token from Settings \u2192 API tokens, sent as the x-access-token header."
 },
 "servers": [
  {
   "url": "https://speakerweave.com"
  }
 ],
 "paths": {
  "/v1/events": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "List Events",
    "description": "List events in the token's organization, newest first.",
    "operationId": "list_events_v1_events_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "pageSize",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Pagesize"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Events V1 Events Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "Get Event",
    "description": "Get one event, returning 404 for an event outside the token's org.",
    "operationId": "get_event_v1_events__event_id__get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Event V1 Events  Event Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/submissions": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "List Submissions",
    "description": "List an event's sessions/submissions with status and track filters.",
    "operationId": "list_submissions_v1_events__event_id__submissions_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "track",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Track"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "pageSize",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Pagesize"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Submissions V1 Events  Event Id  Submissions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "v1-public"
    ],
    "summary": "Create Submission",
    "operationId": "create_submission_v1_events__event_id__submissions_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubmissionCreateRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Submission V1 Events  Event Id  Submissions Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/sessions": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "List Submissions",
    "description": "List an event's sessions/submissions with status and track filters.",
    "operationId": "list_submissions_v1_events__event_id__sessions_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "track",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Track"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "pageSize",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Pagesize"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Submissions V1 Events  Event Id  Sessions Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/submissions/search": {
   "post": {
    "tags": [
     "v1-public"
    ],
    "summary": "Search Submissions",
    "description": "Search submissions using a JSON body compatible with Other Conference/CFP Software.",
    "operationId": "search_submissions_v1_events__event_id__submissions_search_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/SearchRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Search Submissions V1 Events  Event Id  Submissions Search Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/sessions/search": {
   "post": {
    "tags": [
     "v1-public"
    ],
    "summary": "Search Submissions",
    "description": "Search submissions using a JSON body compatible with Other Conference/CFP Software.",
    "operationId": "search_submissions_v1_events__event_id__sessions_search_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/SearchRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Search Submissions V1 Events  Event Id  Sessions Search Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/submissions/{submission_id}": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "Get Submission",
    "operationId": "get_submission_v1_submissions__submission_id__get",
    "parameters": [
     {
      "name": "submission_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Submission Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Submission V1 Submissions  Submission Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "v1-public"
    ],
    "summary": "Update Submission",
    "operationId": "update_submission_v1_submissions__submission_id__patch",
    "parameters": [
     {
      "name": "submission_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Submission Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SubmissionPatchRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Submission V1 Submissions  Submission Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/contacts": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "List Contacts",
    "description": "Legacy Other Conference/CFP Software alias for the event speaker collection.",
    "operationId": "list_contacts_v1_events__event_id__contacts_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "pageSize",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Pagesize"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Contacts V1 Events  Event Id  Contacts Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/contacts/search": {
   "post": {
    "tags": [
     "v1-public"
    ],
    "summary": "Search Contacts",
    "operationId": "search_contacts_v1_events__event_id__contacts_search_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/SearchRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Search Contacts V1 Events  Event Id  Contacts Search Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/speakers": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "List Speakers",
    "operationId": "list_speakers_v1_events__event_id__speakers_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Filter"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "pageSize",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Pagesize"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Speakers V1 Events  Event Id  Speakers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "v1-public"
    ],
    "summary": "Create Speaker",
    "operationId": "create_speaker_v1_events__event_id__speakers_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SpeakerCreateRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Speaker V1 Events  Event Id  Speakers Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/speakers/{speaker_id}": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "Get Speaker",
    "operationId": "get_speaker_v1_speakers__speaker_id__get",
    "parameters": [
     {
      "name": "speaker_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Speaker Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Speaker V1 Speakers  Speaker Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "v1-public"
    ],
    "summary": "Update Speaker",
    "operationId": "update_speaker_v1_speakers__speaker_id__patch",
    "parameters": [
     {
      "name": "speaker_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Speaker Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/routes__v1_routes__SpeakerPatchRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Speaker V1 Speakers  Speaker Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/events/{event_id}/schedule": {
   "get": {
    "tags": [
     "v1-public"
    ],
    "summary": "Get Schedule",
    "operationId": "get_schedule_v1_events__event_id__schedule_get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Event Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Schedule V1 Events  Event Id  Schedule Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SearchRequest": {
    "properties": {
     "status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     },
     "track": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Track"
     },
     "page": {
      "type": "integer",
      "title": "Page",
      "default": 1
     },
     "pageSize": {
      "type": "integer",
      "title": "Pagesize",
      "default": 25
     }
    },
    "type": "object",
    "title": "SearchRequest"
   },
...
  },
  "securitySchemes": {
   "apiToken": {
    "type": "apiKey",
    "in": "header",
    "name": "x-access-token"
   }
  }
 },
 "security": [
  {
   "apiToken": []
  }
 ]
}